requirements.txt
# Discord library
discord.py>=2.3.0

# Database
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Scheduling
APScheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Type checking
mypy>=1.5.0

# Linting
flake8>=6.1.0
black>=23.7.0

# Environment variables
python-dotenv>=1.0.0

# Logging
structlog>=23.1.0

# Utilities
python-dateutil>=2.8.2
